feat: format changed to shortcode
Issue with loading parameters from config file resolved. Due to desire to not require users to
modify the any of hinode's templates, it was decided to implement in shortcode form.

BREAKING CHANGE: This push may restrict the popup to appear only on the pages where it is
implemented, rather than site wide implementation as originally targeted for implementation.